unify solve time data type, create and use a formatter service, prevent screen from turning off when on the timer activity, change the way scrambles are generated (retrieve a string array and format it as needed), retrieve and cache best averages and do a bit of refactoring

+++ datamodel/src/com/cube/nanotimer/vo/SolveAverages.java
@@ -2,80 +2,80 @@ package com.cube.nanotimer.vo;
 
 public class SolveAverages {
 
-  private Float avgOf5;
-  private Float avgOf12;
-  private Float avgOf100;
-  private Float avgOf1000;
+  private Long avgOf5;
+  private Long avgOf12;
+  private Long avgOf100;
+  private Long avgOf1000;
 
-  private Float bestOf5;
-  private Float bestOf12;
-  private Float bestOf100;
-  private Float bestOf1000;
+  private Long bestOf5;
+  private Long bestOf12;
+  private Long bestOf100;
+  private Long bestOf1000;
 
   public SolveAverages() {
   }
 
-  public Float getAvgOf5() {
+  public Long getAvgOf5() {
     return avgOf5;
   }
 
-  public void setAvgOf5(Float avgOf5) {
+  public void setAvgOf5(Long avgOf5) {
     this.avgOf5 = avgOf5;
   }
 
-  public Float getAvgOf12() {
+  public Long getAvgOf12() {
     return avgOf12;
   }
 
-  public void setAvgOf12(Float avgOf12) {
+  public void setAvgOf12(Long avgOf12) {
     this.avgOf12 = avgOf12;
   }
 
-  public Float getAvgOf100() {
+  public Long getAvgOf100() {
     return avgOf100;
   }
 
-  public void setAvgOf100(Float avgOf100) {
+  public void setAvgOf100(Long avgOf100) {
     this.avgOf100 = avgOf100;
   }
 
-  public Float getAvgOf1000() {
+  public Long getAvgOf1000() {
     return avgOf1000;
   }
 
-  public void setAvgOf1000(Float avgOf1000) {
+  public void setAvgOf1000(Long avgOf1000) {
     this.avgOf1000 = avgOf1000;
   }
 
-  public Float getBestOf5() {
+  public Long getBestOf5() {
     return bestOf5;
   }
 
-  public void setBestOf5(Float bestOf5) {
+  public void setBestOf5(Long bestOf5) {
     this.bestOf5 = bestOf5;
   }
 
-  public Float getBestOf12() {
+  public Long getBestOf12() {
     return bestOf12;
   }
 
-  public void setBestOf12(Float bestOf12) {
+  public void setBestOf12(Long bestOf12) {
     this.bestOf12 = bestOf12;
   }
 
-  public Float getBestOf100() {
+  public Long getBestOf100() {
     return bestOf100;
   }
 
-  public void setBestOf100(Float bestOf100) {
+  public void setBestOf100(Long bestOf100) {
     this.bestOf100 = bestOf100;
   }
 
-  public Float getBestOf1000() {
+  public Long getBestOf1000() {
     return bestOf1000;
   }
 
-  public void setBestOf1000(Float bestOf1000) {
+  public void setBestOf1000(Long bestOf1000) {
     this.bestOf1000 = bestOf1000;
   }
 }
